# docs/swagger.yaml
definitions:
  helpers.ApiResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/movies/:
    get:
      consumes:
      - application/json
      description: Get a list of movies with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of movies per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
      summary: Get Movies
      tags:
      - Movies
  /api/movies/create:
    post:
      consumes:
      - application/json
      description: Create Movie Record on movie collection
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
      summary: Create Movie
      tags:
      - Movies
  /api/movies/search:
    post:
      consumes:
      - application/json
      description: Search for movies based on filters
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
      summary: Search Movies
      tags:
      - Movies
  /api/movies/update/:id:
    put:
      consumes:
      - application/json
      description: Update Movie with object id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
      summary: Update Movie
      tags:
      - Movies
  /api/theaters:
    get:
      consumes:
      - application/json
      description: Get a list of theaters with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of theaters per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
      summary: Get Theaters
      tags:
      - Theaters
swagger: "2.0"